pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "seliverstova-valerya",email = "seliverstova.valerya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.5"
sphinx = "^8.2.3"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
httpx = "^0.28.1"
pytest-cov = "^7.0.0"
aiosqlite = "^0.21.0"
fakeredis = "^2.31.3"
sphinx-autobuild = "^2025.8.25"
myst-parser = "^4.0.1"
sphinx-autodoc-typehints = "^3.2.0"
sphinxcontrib-mermaid = "^1.0.0"
sphinx-rtd-theme = "^3.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"

[tool.poetry]
package-mode = false
